requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
playwright
google-genai